+++ src/app/@modal/(.)sn/recipes/[recipeKey]/ModalRecipes.tsx
@@ -1,23 +1,23 @@
 'use client';
-
 import { useRouter } from 'next/navigation';
 
 import style from './style.module.scss';
 
 import { Modal2 } from '@/components/Modal';
-import NewRecipeContainer from '@/app/recipes/new/NewRecipeContainer';
 
-function NewRecipeModal() {
+import RecipePageContainer from '@/app/sn/recipes/[recipeKey]/RecipePageContainer';
+
+function ModalRecipes() {
   const router = useRouter();
   const onDismiss = () => router.back();
 
   return (
-    <Modal2 isOpen={true} onClose={onDismiss} closeOnOutSideClick={false}>
+    <Modal2 isOpen={true} onClose={onDismiss}>
       <div className={style.container}>
-        <NewRecipeContainer />
+        <RecipePageContainer />
       </div>
     </Modal2>
   );
 }
 
-export default NewRecipeModal;
+export default ModalRecipes;
