+++ src/app/@modal/(.)sn/recipes/new/NewRecipeModal.tsx
@@ -1,23 +1,23 @@
 'use client';
+
 import { useRouter } from 'next/navigation';
 
 import style from './style.module.scss';
 
-import RecipePageContainer from '@/app/recipes/[recipeKey]/RecipePageContainer';
-
 import { Modal2 } from '@/components/Modal';
+import NewRecipeContainer from '@/app/sn/recipes/new/NewRecipeContainer';
 
-function ModalRecipes() {
+function NewRecipeModal() {
   const router = useRouter();
   const onDismiss = () => router.back();
 
   return (
-    <Modal2 isOpen={true} onClose={onDismiss}>
+    <Modal2 isOpen={true} onClose={onDismiss} closeOnOutSideClick={false}>
       <div className={style.container}>
-        <RecipePageContainer />
+        <NewRecipeContainer />
       </div>
     </Modal2>
   );
 }
 
-export default ModalRecipes;
+export default NewRecipeModal;
