+++ src/app/components/NewRecipe/components/SearchProductModal/ExistingProduct.tsx
@@ -3,8 +3,7 @@ import { memo } from 'react';
 import style from './style.module.scss';
 
 import { Product } from '@/services/products/type';
-
-import { ProductInfoContent } from '@/app/recipes/[recipeKey]/Recipe/IngredientList/IngredientInformation/Product';
+import { ProductInfoContent } from '@/app/sn/recipes/[recipeKey]/Recipe/IngredientList/IngredientInformation/Product';
 
 import Icon from '@/components/Icon';
 import CustomImage from '@/components/CustomImage';
